error catching when no sample analyzed has hit
If no hit in node list has >10 reads AND declining edit distance, the outputting of the heatmap throws an error. This update checks to make sure at least one sample would be included in the heatmapping output. If not it will print to stdout "No hit with at least a declining edit distance"